refactor(users list): Changed directory location
